Change thread ID type to usize

+++ src/lib.rs
@@ -37,20 +37,20 @@ extern crate kernel32;
 /// number. Calling this function from a different thread will return a
 /// different number.
 #[inline]
-pub fn get() -> u64 {
+pub fn get() -> usize {
     get_internal()
 }
 
 #[cfg(unix)]
 #[inline]
-fn get_internal() -> u64 {
-    unsafe { libc::pthread_self() as u64 }
+fn get_internal() -> usize {
+    unsafe { libc::pthread_self() as usize }
 }
 
 #[cfg(windows)]
 #[inline]
-fn get_internal() -> u64 {
-    unsafe { kernel32::GetCurrentThreadId() as u64 }
+fn get_internal() -> usize {
+    unsafe { kernel32::GetCurrentThreadId() as usize }
 }
 
 #[test]
